# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MDW Panel API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface RegisterInput {
  /** @minLength 3 */
  username: string;
  email: string;
  /** @minLength 6 */
  password: string;
  inviteCode?: string;
}

export interface ChatMessage {
  id: number;
  userId: number;
  username: string;
  message: string;
  createdAt: string;
}

export interface SendMessageInput {
  /**
     * @minLength 1
     * @maxLength 500
     */
  message: string;
}

export interface InviteCodeResponse {
  code: string;
}

export interface InviteCodeInput {
  /** @minLength 1 */
  code: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface User {
  id: number;
  username: string;
  email: string;
  role: UserRole;
  banned?: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface UserUpdate {
  username?: string;
  email?: string;
  role?: UserUpdateRole;
  banned?: boolean;
  password?: string;
}

export interface UserListResponse {
  users: User[];
  total: number;
  page: number;
  limit: number;
}

export type LicenseKeyStatus = typeof LicenseKeyStatus[keyof typeof LicenseKeyStatus];


export const LicenseKeyStatus = {
  active: 'active',
  expired: 'expired',
  banned: 'banned',
} as const;

export interface LicenseKey {
  id: number;
  key: string;
  game: string;
  status: LicenseKeyStatus;
  /** Duration in days */
  duration: number;
  /** @nullable */
  note?: string | null;
  /** @nullable */
  hwid?: string | null;
  /** @nullable */
  lastUsedAt?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  username?: string | null;
  createdAt: string;
}

export interface KeyInput {
  game: string;
  duration: number;
  note?: string;
  /** @nullable */
  userId?: number | null;
}

export type KeyUpdateStatus = typeof KeyUpdateStatus[keyof typeof KeyUpdateStatus];


export const KeyUpdateStatus = {
  active: 'active',
  expired: 'expired',
  banned: 'banned',
} as const;

export interface KeyUpdate {
  status?: KeyUpdateStatus;
  note?: string;
  duration?: number;
  /** @nullable */
  userId?: number | null;
  /** @nullable */
  hwid?: string | null;
}

export interface BulkKeyInput {
  game: string;
  duration: number;
  /**
     * @minimum 1
     * @maximum 100
     */
  count: number;
  note?: string;
}

export interface KeyListResponse {
  keys: LicenseKey[];
  total: number;
  page: number;
  limit: number;
}

export type GameStatus = typeof GameStatus[keyof typeof GameStatus];


export const GameStatus = {
  active: 'active',
  maintenance: 'maintenance',
  disabled: 'disabled',
} as const;

export interface Game {
  id: number;
  name: string;
  slug: string;
  status: GameStatus;
  /** @nullable */
  version?: string | null;
  /** @nullable */
  description?: string | null;
  createdAt: string;
}

export type GameInputStatus = typeof GameInputStatus[keyof typeof GameInputStatus];


export const GameInputStatus = {
  active: 'active',
  maintenance: 'maintenance',
  disabled: 'disabled',
} as const;

export interface GameInput {
  name: string;
  slug: string;
  version?: string;
  description?: string;
  status?: GameInputStatus;
}

export type GameUpdateStatus = typeof GameUpdateStatus[keyof typeof GameUpdateStatus];


export const GameUpdateStatus = {
  active: 'active',
  maintenance: 'maintenance',
  disabled: 'disabled',
} as const;

export interface GameUpdate {
  name?: string;
  version?: string;
  description?: string;
  status?: GameUpdateStatus;
}

export interface DashboardStats {
  totalKeys: number;
  activeKeys: number;
  expiredKeys: number;
  bannedKeys: number;
  totalUsers: number;
  totalGames: number;
  todayConnects: number;
  recentConnects: number;
}

export interface ActivityLog {
  id: number;
  action: string;
  key: string;
  game: string;
  success: boolean;
  /** @nullable */
  ipAddress?: string | null;
  /** @nullable */
  hwid?: string | null;
  /** @nullable */
  reason?: string | null;
  createdAt: string;
}

export type ListKeysParams = {
game?: string;
status?: ListKeysStatus;
page?: number;
limit?: number;
};

export type ListKeysStatus = typeof ListKeysStatus[keyof typeof ListKeysStatus];


export const ListKeysStatus = {
  active: 'active',
  expired: 'expired',
  banned: 'banned',
} as const;

export type ListUsersParams = {
page?: number;
limit?: number;
};

export type GetRecentActivityParams = {
limit?: number;
};

export type GetChatMessagesParams = {
limit?: number;
};