# lib/api-zod/src/generated/types/game.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MDW Panel API
 * OpenAPI spec version: 0.1.0
 */
import type { GameStatus } from './gameStatus';

export interface Game {
  id: number;
  name: string;
  slug: string;
  status: GameStatus;
  /** @nullable */
  version?: string | null;
  /** @nullable */
  description?: string | null;
  createdAt: Date;
}